blackjack
a game in which if the player gets 21 points on the first two cards, it
is called a black jack and the player wins. The goal is to get to this
number as close as possible without going above 21. Implemented the
algorithms behind each move for the player's win and classes for
Blackjack, Card, Dealer, Deck and Player.